OC_OCS_xx to \OC\OCS\xx
* Moved OC_OCS_xx to proper namespace
* OC_OCS_xx is still there for legacy reasons as a wrapper
