+++ lib/private/OCS/Person.php
@@ -1,6 +1,7 @@
 <?php
 /**
- * @author Roeland Jago Douma <rullzer@owncloud.com>
+ * @author Bart Visscher <bartv@thisnet.nl>
+ * @author Morris Jobke <hey@morrisjobke.de>
  * @author Tom Needham <tom@owncloud.com>
  *
  * @copyright Copyright (c) 2016, ownCloud, Inc.
@@ -20,8 +21,23 @@
  *
  */
 
-/**
- * @deprecated Since 9.1.0 use \OC\OCS\Cloud
- */
-class OC_OCS_Cloud extends \OC\OCS\Cloud {
+namespace OC\OCS;
+
+class Person {
+
+	public static function check() {
+		$login = isset($_POST['login']) ? $_POST['login'] : false;
+		$password = isset($_POST['password']) ? $_POST['password'] : false;
+		if($login && $password) {
+			if(\OC_User::checkPassword($login, $password)) {
+				$xml['person']['personid'] = $login;
+				return new \OC_OCS_Result($xml);
+			} else {
+				return new \OC_OCS_Result(null, 102);
+			}
+		} else {
+			return new \OC_OCS_Result(null, 101);
+		}
+	}
+
 }
