+++ lib/private/legacy/ocs/person.php
@@ -1,7 +1,6 @@
 <?php
 /**
- * @author Bart Visscher <bartv@thisnet.nl>
- * @author Morris Jobke <hey@morrisjobke.de>
+ * @author Roeland Jago Douma <rullzer@owncloud.com>
  * @author Tom Needham <tom@owncloud.com>
  *
  * @copyright Copyright (c) 2016, ownCloud, Inc.
@@ -21,21 +20,8 @@
  *
  */
 
-class OC_OCS_Person {
-
-	public static function check() {
-		$login = isset($_POST['login']) ? $_POST['login'] : false;
-		$password = isset($_POST['password']) ? $_POST['password'] : false;
-		if($login && $password) {
-			if(OC_User::checkPassword($login, $password)) {
-				$xml['person']['personid'] = $login;
-				return new OC_OCS_Result($xml);
-			} else {
-				return new OC_OCS_Result(null, 102);
-			}
-		} else {
-			return new OC_OCS_Result(null, 101);
-		}
-	}
-
+/**
+ * @deprecated Since 9.1.0 use \OC\OCS\Person
+ */
+class OC_OCS_Person extends \OC\OCS\Person {
 }
